requirements.txt
############################################################
# HTTP clients & web server
############################################################
requests==2.32.3
aiohttp>=3.9.0

############################################################
# Logging / env
############################################################
structlog==23.2.0
python-dotenv==1.0.1

############################################################
# Azure SDKs
############################################################
azure-cosmos==4.6.0

############################################################
# LLM / Agent ecosystem
############################################################
langchain
langchain-core
langchain-openai

############################################################
# Bot Framework SDK
############################################################
botbuilder-integration-aiohttp>=4.15.0
